NOT WORKING VERSION; Separate Presentation Layer from the Logic Layer
